Update to use coroutines
Updated to use coroutines all the way in place of eventbus
