+++ fluxc/src/main/java/org/wordpress/android/fluxc/store/SiteStore.kt
@@ -38,7 +38,6 @@ import org.wordpress.android.fluxc.action.SiteAction.FETCHED_USER_ROLES
 import org.wordpress.android.fluxc.action.SiteAction.FETCHED_WPCOM_SITE_BY_URL
 import org.wordpress.android.fluxc.action.SiteAction.FETCH_BLOCK_LAYOUTS
 import org.wordpress.android.fluxc.action.SiteAction.FETCH_CONNECT_SITE_INFO
-import org.wordpress.android.fluxc.action.SiteAction.FETCH_DOMAINS
 import org.wordpress.android.fluxc.action.SiteAction.FETCH_DOMAIN_SUPPORTED_COUNTRIES
 import org.wordpress.android.fluxc.action.SiteAction.FETCH_DOMAIN_SUPPORTED_STATES
 import org.wordpress.android.fluxc.action.SiteAction.FETCH_JETPACK_CAPABILITIES
@@ -72,6 +71,8 @@ import org.wordpress.android.fluxc.model.RoleModel
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.model.SitesModel
 import org.wordpress.android.fluxc.network.BaseRequest.BaseNetworkError
+import org.wordpress.android.fluxc.network.rest.wpcom.WPComGsonRequestBuilder.Response.Error
+import org.wordpress.android.fluxc.network.rest.wpcom.WPComGsonRequestBuilder.Response.Success
 import org.wordpress.android.fluxc.network.rest.wpcom.site.Domain
 import org.wordpress.android.fluxc.network.rest.wpcom.site.DomainSuggestionResponse
 import org.wordpress.android.fluxc.network.rest.wpcom.site.GutenbergLayout
@@ -99,6 +100,8 @@ import org.wordpress.android.fluxc.store.SiteStore.DomainSupportedStatesErrorTyp
 import org.wordpress.android.fluxc.store.SiteStore.ExportSiteErrorType.GENERIC_ERROR
 import org.wordpress.android.fluxc.store.SiteStore.PlansErrorType.NOT_AVAILABLE
 import org.wordpress.android.fluxc.store.SiteStore.SiteErrorType.DUPLICATE_SITE
+import org.wordpress.android.fluxc.store.SiteStore.SiteErrorType.UNAUTHORIZED
+import org.wordpress.android.fluxc.store.SiteStore.SiteErrorType.UNKNOWN_SITE
 import org.wordpress.android.fluxc.tools.CoroutineEngine
 import org.wordpress.android.fluxc.utils.SiteErrorUtils
 import org.wordpress.android.util.AppLog
@@ -652,19 +655,6 @@ open class SiteStore
         @JvmField val domains: List<Domain>? = null
     ) : Payload<SiteError>() {
         constructor(site: SiteModel, error: SiteError) : this(site) {
-            this.error = error
-        }
-    }
-
-    data class OnSiteDomainsFetched(
-        @JvmField val site: SiteModel,
-        @JvmField val domains: List<Domain>?
-    ) : OnChanged<SiteError>() {
-        constructor(
-            site: SiteModel,
-            domains: List<Domain>?,
-            error: SiteError?
-        ) : this(site, domains) {
             this.error = error
         }
     }
@@ -1262,9 +1252,6 @@ open class SiteStore
             FETCHED_JETPACK_CAPABILITIES -> handleFetchedJetpackCapabilities(
                     action.payload as FetchedJetpackCapabilitiesPayload
             )
-            FETCH_DOMAINS -> coroutineEngine.launch(T.MAIN, this, "Fetch site domains") {
-                emitChange(fetchSiteDomains(action.payload as SiteModel))
-            }
         }
     }
 
@@ -1844,8 +1831,22 @@ open class SiteStore
         emitChange(event)
     }
 
-    suspend fun fetchSiteDomains(siteModel: SiteModel): OnSiteDomainsFetched {
-        val result = siteRestClient.fetchSiteDomains(siteModel)
-        return OnSiteDomainsFetched(siteModel, result.domains, result.error)
-    }
+    suspend fun fetchSiteDomains(siteModel: SiteModel): FetchedDomainsPayload =
+            coroutineEngine.withDefaultContext(T.API, this, "Fetch site domains") {
+                return@withDefaultContext when (val response =
+                        siteRestClient.fetchSiteDomains(siteModel)) {
+                            is Success -> {
+                                FetchedDomainsPayload(siteModel, response.data.domains)
+                            }
+                            is Error -> {
+                                val siteErrorType = when (response.error.apiError) {
+                                    "unauthorized" -> UNAUTHORIZED
+                                    "unknown_blog" -> UNKNOWN_SITE
+                                    else -> SiteErrorType.GENERIC_ERROR
+                                }
+                                val domainsError = SiteError(siteErrorType, response.error.message)
+                                FetchedDomainsPayload(siteModel, domainsError)
+                            }
+                        }
+            }
 }
